Bump version.
This version drops the often problematic attempt to add binaries to the PATH
and, instead, is always invoked as erbtex but with the option to specify the TeX
variant to run on the resulting code with an --invoke=<tex_program> command-line
option. The <tex_program> can be something like 'pdflatex' (the default) or
'tex', etc.